+++ src/NJsonSchema/JsonSchema4.cs
@@ -458,7 +458,20 @@ namespace NJsonSchema
 
         /// <summary>Gets the xml object of the type. </summary>
         [JsonProperty("xml", DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        public JsonXmlObject Xml { get; set; }
+        public JsonXmlObject Xml {
+            get
+            {
+                return _xmlObject;
+            }
+            set
+            {
+                _xmlObject = value;
+                //_xmlObject.ParentSchema = this;
+            }
+        }
+
+        [JsonIgnore]
+        private JsonXmlObject _xmlObject;
 
         /// <summary>Gets the pattern properties of the type. </summary>
         [JsonIgnore]
